Finish draft implementation of FastUpdateManager functionality.

+++ contracts/fastUpdates/implementation/FastUpdaters.sol
@@ -4,8 +4,8 @@ pragma solidity 0.8.18;
 import { ECPoint } from "../lib/Sortition.sol";
 
 contract FastUpdaters {
-    function numParticipants() public returns (uint) {
-        
+    function numParticipants() public returns (uint8) {
+
     }
 
     function sortitionPublicKey(address provider) public returns (ECPoint memory) {
